Html resolving method extraction
