+++ .npmignore
@@ -12,8 +12,8 @@
 !typeNodes.js
 !decorators/languageVariantsDecorator.js
 !decorators/typeItemDecorator.js
-!decorators/inlineLinkedItemsDecorator.js
-!decorators/linkedItemsDecorator.js
+!decorators/richTextElementDecorator.js
+!decorators/linkedItemsElementDecorator.js
 
 !package.json
 !.npmignore
