+++ src/decorators/linkedItemsElementDecorator.js
